Lukas created new RPG style testing map, IMGui is finally here, music, more animations and fixing warnings

+++ src/Level.cpp
@@ -3,13 +3,13 @@
 namespace Vigilant {
 
     void Level::render() {
-        for (int i = 0; i < layers.size(); i++) {
+        for (size_t i = 0; i < layers.size(); i++) {
             layers[i]->render();
         }
     }
 
     void Level::update() {
-        for (int i = 0; i < layers.size(); i++) {
+        for (size_t i = 0; i < layers.size(); i++) {
             layers[i]->update();
         }
     }
